+++ alembic/versions/d4e5f6a7b8c4_create_copilot_audit_log.py
@@ -18,8 +18,8 @@ def upgrade() -> None:
     op.create_table(
         'copilot_audit_log',
         sa.Column('id', sa.UUID(), primary_key=True, server_default=sa.text('gen_random_uuid()')),
-        sa.Column('company_id', sa.Integer(), sa.ForeignKey('companies.id'), nullable=False),
-        sa.Column('user_id', sa.Integer(), sa.ForeignKey('users.id'), nullable=False),
+        sa.Column('company_id', sa.BigInteger(), sa.ForeignKey('companies.id'), nullable=False),
+        sa.Column('user_id', sa.BigInteger(), sa.ForeignKey('users.id'), nullable=False),
         sa.Column('conversation_id', sa.UUID(), nullable=False),
         sa.Column('plan_id', sa.UUID(), nullable=False),
         sa.Column('step_id', sa.Text(), nullable=False),
